requirements.txt
customtkinter
opencv-python
mediapipe
numpy
pandas
plyer
pillow